fix chats gql
