+++ src/features/chats/episodes/api/useMutateChatEpisode.ts
@@ -1,21 +1,21 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { insertChatDocument } from "src/documents/chats";
-
-import { PageParams } from "src/features/chats/types";
+import { PageParams } from "src/features/chats/common/types";
 import {
-  GetChatsWorkQuery,
+  GetChatsEpisodeQuery,
   InsertChatMutation,
   InsertChatMutationVariables,
 } from "src/gql/graphql";
 import { Error } from "src/libs/error";
+
 import { client } from "src/libs/graphqlClient";
 
 type PrevData = {
-  pages: GetChatsWorkQuery[];
+  pages: GetChatsEpisodeQuery[];
   pageParams: PageParams[];
 };
 
-export const useMutateChatWork = () => {
+export const useMutateChatEpisode = () => {
   const queryClient = useQueryClient();
   const insertChat = useMutation<
     InsertChatMutation,
@@ -26,13 +26,11 @@ export const useMutateChatWork = () => {
     onSuccess: (data) => {
       if (!data.insert_chats_one) return;
 
-      const { work_id, comment_time } = data.insert_chats_one;
-
-      if (!work_id) return;
+      const { episode_id, comment_time } = data.insert_chats_one;
 
       const prevData = queryClient.getQueryData<PrevData>([
         "chats",
-        { work_id },
+        { episode_id },
       ]);
 
       if (!prevData) return;
@@ -42,30 +40,30 @@ export const useMutateChatWork = () => {
       const newPages = prevData.pages.map((page, index) => {
         if (index !== mutateCommentPageIndex || !data.insert_chats_one)
           return page;
-        const nextTimeIndex = page.chats_by_work_id.findIndex(
+        const nextTimeIndex = page.chats_by_episode_id.findIndex(
           (comment) => comment.comment_time >= comment_time
         );
 
-        const prevEpisodes: GetChatsWorkQuery = {
-          chats_by_work_id: [...page.chats_by_work_id],
+        const prevEpisodes: GetChatsEpisodeQuery = {
+          chats_by_episode_id: [...page.chats_by_episode_id],
         };
 
         if (nextTimeIndex === -1) {
-          prevEpisodes.chats_by_work_id.push({
+          prevEpisodes.chats_by_episode_id.push({
             ...data.insert_chats_one,
           });
 
           return prevEpisodes;
         }
 
-        prevEpisodes.chats_by_work_id.splice(nextTimeIndex, 0, {
+        prevEpisodes.chats_by_episode_id.splice(nextTimeIndex, 0, {
           ...data.insert_chats_one,
         });
 
         return prevEpisodes;
       });
 
-      queryClient.setQueryData(["chats", { work_id }], {
+      queryClient.setQueryData(["chats", { episode_id }], {
         pages: newPages,
         pageParams: prevData.pageParams,
       });
