+++ src/features/chats/works/hooks/useSubmitWork.ts
@@ -1,21 +1,19 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
 import { useState } from "react";
 import { useNotificationState } from "src/components/Notification/store";
-import { useMutateChatEpisode } from "src/features/chats/api/useMutateChatEpisode";
-import { useSubmitChat } from "src/features/chats/hooks/useSubmitChat";
+import { useSubmitChat } from "src/features/chats/common/hooks/useSubmitChat";
+import { useMutateChatWork } from "src/features/chats/works/api/useMutateChatWork";
 import { useGlobalState } from "src/store/global/globalStore";
 import { getIp } from "src/utils/client/getIp";
 
 type Args = {
-  episode_id: string;
+  work_id: number;
 };
 
-export const useSubmitChatEpisode = ({ episode_id }: Args) => {
-  const { insertChat } = useMutateChatEpisode();
+export const useSubmitWork = ({ work_id }: Args) => {
+  const { insertChat } = useMutateChatWork();
   const { setValue, user, getTime, value } = useSubmitChat();
-  const onNotification = useNotificationState((state) => state.onShow);
   const authLoading = useGlobalState((state) => state.authLoading);
+  const onNotification = useNotificationState((state) => state.onShow);
   const [ipLoading, setIpLoading] = useState(false);
 
   const onSubmitHandler = async (e: React.FormEvent<HTMLFormElement>) => {
@@ -23,11 +21,10 @@ export const useSubmitChatEpisode = ({ episode_id }: Args) => {
     try {
       if (!value.trim() || authLoading) return;
       setIpLoading(true);
-
       const ip = await getIp();
 
       const object = {
-        episode_id,
+        work_id,
         content: value,
         comment_time: getTime(),
         commenter_name: user?.user_name || "匿名",
@@ -50,7 +47,7 @@ export const useSubmitChatEpisode = ({ episode_id }: Args) => {
   return {
     onSubmitHandler,
     isLoading: insertChat.isPending || ipLoading,
-    value,
     setValue,
+    value,
   };
 };
