+++ src/features/episodes/components/index.tsx
@@ -7,8 +7,8 @@ import { Loader } from "src/components/Elements/Loader";
 import { Skeleton } from "src/components/Elements/Skeleton";
 import { ErrorMessage } from "src/components/Error/items/ErrorMessage";
 import { TimerModal } from "src/components/Modal/Timer";
-import { EpisodeChatInput } from "src/features/chats/components/EpisodeChatInput";
-import { EpisodeChats } from "src/features/chats/components/EpisodeChats";
+import { EpisodeChatInput } from "src/features/chats/episodes/components/EpisodeChatInput";
+import { EpisodeChats } from "src/features/chats/episodes/components/EpisodeChats";
 import { EpisodeCommentInput } from "src/features/comments/components/EpisodeCommentInput";
 import { EpisodeComments } from "src/features/comments/components/EpisodeComments";
 import { useEpisode } from "src/features/episodes/hooks/useEpisode";
